Have (Epoll|KQueue)RecvByteAllocatorHandle extend DelegatingHandle (#9060)
Motivation
These implementations delegate most of their methods to an existing Handle and previously extended
RecvByteBufAllocator.DelegatingHandle. This was reverted in (#6322) with the introduction of
ExtendedHandle but it's not clear to me why it needed to be - the code looks a lot cleaner without
the boilerplate.

Modifications
Have (Epoll|KQueue)RecvByteAllocatorHandle extend DelegatingHandle again, while still implementing
ExtendedHandle.

Result
68 fewer lines.