DnsNameResolver.resolveAll(...) should also contain non preferred addresses (#9044)
Motivation:
At the moment we basically drop all non prefered addresses when calling DnsNameResolver.resolveAll(...).
This is just incorrect and was introduced by 4cd39cc. More correct is to still retain these but sort the
returned List to have the prefered addresses on the beginning of the List. This also ensures resolve(...)
will return the correct return type.

Modifications:
Introduce PreferredAddressTypeComperator which we use to sort the List so it will contain the preferred
address type first.  Add unit test to verify behaviour

Result:
Include not only preferred addresses in the List that is returned by resolveAll(...)